+++ src/components/template-import-export/index.ts
@@ -1,4 +1,4 @@
 export * from './routes'
 export * from './updateHashes'
 export * from './getTemplateLinkedEntities'
-export * from './commitTemplate'
+export * from './operations/commitTemplate'
