+++ src/components/template-import-export/operations/exportTemplate.ts
@@ -1,13 +1,13 @@
 import path from 'path'
 import fsx from 'fs-extra'
-import { ApiError } from './ApiError'
-import db from './databaseMethods'
-import { FILES_FOLDER, FILES_TEMP_FOLDER } from '../../constants'
+import { ApiError } from '../../../ApiError'
+import db from '../databaseMethods'
+import { FILES_FOLDER, FILES_TEMP_FOLDER } from '../../../constants'
 import { DateTime } from 'luxon'
-import config from '../../config'
+import config from '../../../config'
 import archiver from 'archiver'
-import { buildTemplateStructure } from './buildTemplateStructure'
-import { PgTemplate } from './types'
+import { buildTemplateStructure } from '../buildTemplateStructure'
+import { PgTemplate } from '../types'
 
 export const exportTemplate = async (templateId: number) => {
   console.log(`Exporting template: ${templateId}...`)
