+++ src/components/template-import-export/operations/index.ts
@@ -1,0 +1,5 @@
+export * from './checkTemplate'
+export * from './commitTemplate'
+export * from './duplicateTemplate'
+export * from './exportTemplate'
+export * from './importTemplate'
